+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [2.0.3] - [Unreleased]
+### Fixed
+- Improve automatic checkmarks for monthly habits (@iSoron, 947)
+- Fix small theme issues (@iSoron)
+- Fix ANR on some Samsung phones (@iSoron, #962)
+
 ## [2.0.2] - 2021-05-23
 
 ### Changed